Add handling errors of GET, PATCH and DELETE requests: replace custom exceptions with existing ones, move all exception handling to GlobalExceptionHandler class, move exception throwing to service level where possible

+++ src/main/java/com/dolgikh/scriptorium/services/AuthorsService.java
@@ -2,16 +2,22 @@ package com.dolgikh.scriptorium.services;
 
 import com.dolgikh.scriptorium.models.Author;
 import com.dolgikh.scriptorium.repositories.AuthorsRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional(readOnly = true)
 public class AuthorsService {
     private final AuthorsRepository authorsRepository;
+
+    private String notFoundMessage(int id) {
+        return "Author with id " + id + " was not found";
+    }
 
     @Autowired
     public AuthorsService(AuthorsRepository authorsRepository) {
@@ -23,7 +29,12 @@ public class AuthorsService {
     }
 
     public Author findOne(int id) {
-        return authorsRepository.findById(id).orElse(null);
+        Optional<Author> author = authorsRepository.findById(id);
+
+        if (author.isEmpty())
+            throw new EntityNotFoundException(notFoundMessage(id));
+
+        return author.get();
     }
 
     @Transactional
@@ -32,7 +43,19 @@ public class AuthorsService {
     }
 
     @Transactional
+    public void update(Author author, int id) {
+        if (authorsRepository.findById(id).isEmpty())
+            throw new EntityNotFoundException(notFoundMessage(id));
+
+        author.setId(id);
+        authorsRepository.save(author);
+    }
+
+    @Transactional
     public void delete(int id) {
+        if (authorsRepository.findById(id).isEmpty())
+            throw new EntityNotFoundException(notFoundMessage(id));
+
         authorsRepository.deleteById(id);
     }
 }
